requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
opencv-python>=4.8.0
streamlit-chat>=0.1.0